# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ECHO Mind Path A API",
    "description": "心理健康记录、筛查提示、审核练习、人工接管与审计。非诊断、非紧急服务替代。",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/escalations": {
      "post": {
        "summary": "Create Escalation",
        "description": "用户/服务端创建升级事件（幂等：同 tenant+event_id 返回既有记录）。\n\n用户侧调用链（v0.6.1 客户端闭环）：Android Outbox → POST /v1/escalations；\ndelivery_confirmed_at 在服务端接收时即写入（送达确认，不等于人工已收到）。",
        "operationId": "create_escalation_v1_escalations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Escalations",
        "description": "机构工作台队列：cursor 分页 + 过滤（v0.6.1）。\n\n**契约保持**：响应体仍为数组（与 v0.6 一致）；分页元数据放在响应头\n``X-Next-Cursor``（还有更多页时非空）与 ``X-Total-Filtered``。\n不变量：本端点仅 workbench 角色；evidence_summary 仅持有 step-up 声明时可见。",
        "operationId": "list_escalations_v1_escalations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "opaque cursor（X-Next-Cursor 响应头回填）",
              "title": "Cursor"
            },
            "description": "opaque cursor（X-Next-Cursor 响应头回填）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/escalations/sla-scan": {
      "post": {
        "summary": "Sla Scan",
        "description": "运维定时触发的 SLA 扫描：推进未确认红色事件的自动升级链路（幂等）。",
        "operationId": "sla_scan_v1_escalations_sla_scan_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/escalations/{escalation_id}": {
      "get": {
        "summary": "Escalation Detail",
        "operationId": "escalation_detail_v1_escalations__escalation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Escalation",
        "operationId": "delete_escalation_v1_escalations__escalation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Escalation",
        "operationId": "patch_escalation_v1_escalations__escalation_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/escalations/{escalation_id}/user-status": {
      "get": {
        "summary": "Escalation User Status",
        "description": "用户侧状态查询：只暴露送达确认与接管状态，不暴露内部升级细节。\n\nhuman_acknowledged 仅由显式 ack/takeover 决定；系统通知（notified_*）与\n送达确认（delivery_confirmed）都不得被当作\"人工已收到/已接管\"。",
        "operationId": "escalation_user_status_v1_escalations__escalation_id__user_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/escalations/{escalation_id}/case-review": {
      "get": {
        "summary": "Escalation Case Review",
        "description": "个案复核证据链视图（仅复核类角色 + step-up）。",
        "operationId": "escalation_case_review_v1_escalations__escalation_id__case_review_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/escalations/{escalation_id}/ack": {
      "post": {
        "summary": "Ack Escalation",
        "operationId": "ack_escalation_v1_escalations__escalation_id__ack_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "escalation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Escalation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EscalationCreate": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "title": "Event Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "level": {
            "type": "string",
            "enum": [
              "L2",
              "L3"
            ],
            "title": "Level",
            "default": "L3"
          },
          "trigger": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Trigger"
          },
          "evidence_summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Evidence Summary"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "user_id",
          "trigger",
          "evidence_summary"
        ],
        "title": "EscalationCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}